fix(clean-search): new search module
