+++ backend/airweave/api/v1/api.py
@@ -16,6 +16,7 @@ from airweave.api.v1.endpoints import (
     file_retrieval,
     health,
     organizations,
+    search,
     source_connections,
     sources,
     sync,
@@ -40,6 +41,7 @@ api_router.include_router(
 )
 api_router.include_router(auth_providers.router, prefix="/auth-providers", tags=["auth-providers"])
 api_router.include_router(collections.router, prefix="/collections", tags=["collections"])
+api_router.include_router(search.router, prefix="/collections", tags=["collections"])
 api_router.include_router(connections.router, prefix="/connections", tags=["connections"])
 api_router.include_router(
     source_connections.router, prefix="/source-connections", tags=["source-connections"]
