+++ backend/airweave/search/operations/embed_query.py
@@ -1,0 +1,9 @@
+"""Embed the user query."""
+
+from ._base import SearchOperation
+
+
+class EmbedQuery(SearchOperation):
+    """Embed the user query."""
+
+    pass
