+++ backend/airweave/search/operations/generate_answer.py
@@ -1,0 +1,9 @@
+"""Generate an answer to the user's query."""
+
+from ._base import SearchOperation
+
+
+class GenerateAnswer(SearchOperation):
+    """Generate an answer to the user's query."""
+
+    pass
