+++ backend/airweave/search/operations/retrieval.py
@@ -1,0 +1,9 @@
+"""Retrieve the search results."""
+
+from ._base import SearchOperation
+
+
+class Retrieval(SearchOperation):
+    """Retrieve the search results."""
+
+    pass
