+++ backend/airweave/search/operations/temporal_relevance.py
@@ -1,0 +1,9 @@
+"""Temporal relevance."""
+
+from ._base import SearchOperation
+
+
+class TemporalRelevance(SearchOperation):
+    """Temporal relevance."""
+
+    pass
