+++ backend/airweave/search/operations/user_filter.py
@@ -1,0 +1,9 @@
+"""User-defined Qdrant filter."""
+
+from ._base import SearchOperation
+
+
+class UserFilter(SearchOperation):
+    """User-defined Qdrant filter."""
+
+    pass
