+++ backend/airweave/search/operations2/completion_openai.py
