+++ backend/airweave/search/providers/cohere.py
@@ -1,0 +1,1 @@
+# not implemented error for text generation method
